+++ src/app/fav/page.tsx
@@ -1,67 +1,29 @@
 import Link from "next/link";
 
-import { Selectable } from "kysely";
 import { getServerSession } from "next-auth";
 import { TbMenu, TbUserCircle } from "react-icons/tb";
 
 import { HorizonalSmallChefList } from "@/components/horizontal-small-chef-list/horizontal-small-chef-list";
 import { Login } from "@/components/login";
-import { HorizontalRecipeList, VerticalRecipeList } from "@/features/recipes";
+import { getFavoriteChefs } from "@/features/chefs";
+import {
+  getFavoriteRecipeWithFavoriteCountByUserId,
+  getRecipeWithFavoriteCountByUserId,
+  HorizontalRecipeList,
+  VerticalRecipeList,
+} from "@/features/recipes";
 import { authOptions } from "@/lib/auth";
-import { db } from "@/lib/kysely";
-import { RecipeFavorite } from "@/types/db";
 
 export default async function Page() {
   const session = await getServerSession(authOptions);
-
   const userId = session?.user?.id as string;
-  // お気に入りシェフ
-  const followedChefs = await db.selectFrom("UserFollow").selectAll().where("followerUserId", "=", userId).execute();
-  const followedChefsArray = followedChefs.map((chef) => chef.followedUserId);
-  const chefs = await db.selectFrom("User").selectAll().where("id", "in", followedChefsArray).execute();
 
-  // 新着レシピ
-  const recipeList = await db
-    .selectFrom("Recipe")
-    .innerJoin("RecipeImage", "RecipeImage.recipeId", "Recipe.id")
-    .select([
-      "Recipe.id as id",
-      "Recipe.userId as userId",
-      "Recipe.name as name",
-      "Recipe.description as description",
-      "Recipe.servings as servings",
-      "Recipe.isPublic as isPublic",
-      "Recipe.createdAt as createdAt",
-      "Recipe.updatedAt as updatedAt",
-      "RecipeImage.imgSrc as imgSrc",
-    ])
-    .where("Recipe.isPublic", "=", 1)
-    .where("Recipe.deletedAt", "is", null)
-    .where("RecipeImage.deletedAt", "is", null)
-    .where("userId", "in", followedChefsArray)
-    .orderBy("Recipe.createdAt", "desc")
-    .execute();
+  const chefs = await getFavoriteChefs(userId);
 
-  const recipeIds = recipeList.map((recipe) => recipe["id"]);
+  const followedChefsArray = chefs.map((chef) => chef.id);
+  const recentRecipeList = await getRecipeWithFavoriteCountByUserId(followedChefsArray);
 
-  const recipeFavorites: Pick<Selectable<RecipeFavorite>, "recipeId">[] = await db
-    .selectFrom("RecipeFavorite")
-    .select(["recipeId"])
-    .where("recipeId", "in", recipeIds)
-    .where("deletedAt", "is", null)
-    .execute();
-
-  const recipeFavoriteCounts = recipeFavorites.reduce(function (prev: { [key: string]: number }, current) {
-    prev[current["recipeId"]] = (prev[current["recipeId"]] || 0) + 1;
-    return prev;
-  }, {});
-
-  const recentRecipeList = recipeList.flatMap((recipe) => {
-    return {
-      ...recipe,
-      favoriteCount: recipeFavoriteCounts[recipe.id] ? recipeFavoriteCounts[recipe.id] : 0,
-    };
-  });
+  const favoriteRecipeList = await getFavoriteRecipeWithFavoriteCountByUserId(userId);
 
   return (
     <main className="min-h-screen w-full text-mauve-12">
@@ -95,7 +57,7 @@ export default async function Page() {
           <div className="mb-[10px] ml-4 mt-12">
             <h2 className="text-xl font-bold text-mauve-12">お気に入りレシピ</h2>
           </div>
-          <div className="mx-4">{<VerticalRecipeList recipeList={recentRecipeList} />}</div>
+          <div className="mx-4">{<VerticalRecipeList recipeList={favoriteRecipeList} />}</div>
         </div>
       ) : (
         <Login imgSrc="/images/fav-login.png" />
