+++ src/features/chefs/index.tsx
@@ -2,5 +2,8 @@
 export * from "./components/chef-list/vertical-chef-list-item";
 export * from "./components/chef-list/vertical-chef-list";
 
+// lib
+export * from "./lib/actions";
+
 //type
 export * from "./types";
